refactor: improve code organization and fix imports - Move main.rs to bin directory, remove deprecated MCP client, fix import conflicts, update RiskFactors imports, clean up unused imports, add documentation

+++ src/risk_model.rs
@@ -1,18 +1,18 @@
 use ndarray::{Array2, ArrayD};
 use async_trait::async_trait;
 use crate::{ModelError, MarketData};
-use serde::Serialize;
+use serde::{Serialize, Deserialize};
 
 /// Learned risk factors and their properties.
 /// 
 /// This struct represents the output of the risk model, containing
 /// both the learned risk factors and their covariance matrix.
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RiskFactors {
     /// Risk factor values (N x K matrix where K is number of factors)
-    pub factors: ArrayD<f64>,
+    pub factors: ArrayD<f32>,
     /// Factor returns (regression coefficients)
-    pub covariance: Array2<f64>,
+    pub covariance: Array2<f32>,
 }
 
 /// Core trait for risk factor models.
@@ -45,6 +45,6 @@ pub trait RiskModel: Send + Sync {
     /// * `data` - Market data to use for estimation
     /// 
     /// # Returns
-    /// * `Result<Array2<f64>, ModelError>` - Estimated covariance matrix or error
-    async fn estimate_covariance(&self, data: &MarketData) -> Result<Array2<f64>, ModelError>;
+    /// * `Result<Array2<f32>, ModelError>` - Estimated covariance matrix or error
+    async fn estimate_covariance(&self, data: &MarketData) -> Result<Array2<f32>, ModelError>;
 }
